Add a patch for GTK+3 to enable text input with third party input methods.
If a third party input method such as GoogleIME is enabled, GdkQuartz will
breaks all keymaps. The main cause of this issue is that GdkQuatz tries to
update keymaps on every key event. When you apply this patch, GdkQuartz
updates keymaps only when it is needed. It uses NSNoticationCenter to
observe switching input methods.

I'll report it to upstream after verify the patch.